Increase UVLO turn-on to 43 V
The initial design assumed that the UVLO was a feature of the LM3409 not needed
by this board. It was set at the datasheet values of 10.1 V turn-on with 1.1 V
of hysteresis. However, during testing, I have found that the supply will sink
a maximum of around 600 mA into the load when operated into dropout. The LM3409
datasheet confirms that this is expected behaviour. Since the supply is
designed for LEDs modules with a forward voltage of 37 V, this patch changes
the UVLO to just over 43 volts, meaning that we do not run the LEDs above the
supply's design current of 450 mA.